debug: fix inconsistent behaviour of 'Run to Cursor'
fixes #8731

+++ src/vs/workbench/parts/debug/electron-browser/debugActions.ts
@@ -604,8 +604,10 @@ export class RunToCursorAction extends EditorAction {
 		const oneTimeListener = this.debugService.getActiveSession().onDidEvent(event => {
 			if (event.event === 'stopped' || event.event === 'exit') {
 				const toRemove = this.debugService.getModel().getBreakpoints()
-					.filter(bp => bp.lineNumber === lineNumber && bp.source.uri.toString() === uri.toString()).pop();
-				this.debugService.removeBreakpoints(toRemove.getId());
+					.filter(bp => bp.desiredLineNumber === lineNumber && bp.source.uri.toString() === uri.toString()).pop();
+				if (toRemove) {
+					this.debugService.removeBreakpoints(toRemove.getId());
+				}
 				oneTimeListener.dispose();
 			}
 		});
